Fix area navigation and panel appearances.
