--- app/components/area-directory.js
@@ -1,10 +1,0 @@
-import Component from '@ember/component';
-
-export default Component.extend({
-        
-    actions: { 
-        selectArea(area) {
-            this.get('onSelected')(area.name);
-        }
-    }
-});
