+++ app/templates/char.hbs
@@ -74,7 +74,7 @@
 {{#if showFiles}}
     <div class="panel panel-default">
         
-        <div class="panel panel-body">
+        <div class="panel-body">
             
         <p><b>Existing Character Files:</b></p>
