+++ app/templates/components/file-uploader.hbs
@@ -1,9 +1,9 @@
 <div class="panel panel-default">
-    <div class="panel panel-heading">
+    <div class="panel-heading">
         Upload Files
     </div>
 
-    <div class="panel panel-body">
+    <div class="panel-body">
 
         {{#if showSelector}}
         <p>Select files for upload...</p>
