+++ app/templates/forum-topic.hbs
@@ -31,7 +31,7 @@
 {{#if model.can_reply}}
 
 <div class="panel panel-default">
-    <div class="panel panel-heading">
+    <div class="panel-heading">
         <a data-toggle="collapse" href="#reply">Click to Reply</a>
         
     </div>
