+++ app/templates/locations.hbs
@@ -3,80 +3,21 @@
 
 <p>The locations directory allows you to browse locations (aka 'rooms') on the game, organized by 'area'.</p>
 
-<h2>Areas</h2>
-
-<p class="hint">Select an area to view its details.</p>
+<h2>Area Directory</h2>
 
   <ul>
   {{#each model.directory as |a| }}
-      {{area-directory area=a onSelected=(action "selectArea")}}
+      {{area-directory area=a}}
   {{/each}}
   </ul>
 
-{{#if currentArea}}
-
-<div class="panel panel-default">
-    <div class="panel panel-body">
-
-        <h2>{{currentArea.name}}</h2>
-        {{#if currentArea.parent}}
-        <p><b>Parent Area:</b> <a href="#" {{action (action "selectArea" currentArea.parent)}}>{{currentArea.parent}}</a></p>
-        {{/if}}
-        
-        <div>
-            {{{ansi-format text=currentArea.description}}}
-        </div>
-            <h3>Areas in {{currentArea.name}}</h3>
-             <ul> 
-                {{#each currentArea.children as |c|}}
-                  <li><a href="#" {{action (action "selectArea" c.name)}}>{{c.name}}</a></li>
-                {{/each}}
-            </ul>
-            
-            <h3>Locations in {{currentArea.name}}</h3>
-             <ul> 
-                {{#each currentArea.locations as |loc|}}
-                  <li>{{#link-to 'location' loc.id}}{{loc.name}}{{/link-to}}</li>
-                {{/each}}
-            </ul>
-
-        <div class="row">
-            <div class="col col-xs-12">
-                <div class="pull-right">
-            
-            
-                {{#if model.can_manage}}
-                  <button {{action (mut confirmDelete) true}} class="btn btn-warning">Delete</button>
-                  {{#link-to 'area-edit' currentArea.id class="btn btn-primary"}}Edit{{/link-to}}
-                {{/if}}
-              </div>
-          </div>
-        </div>        
-    </div>
-</div>
-
-
-{{/if}}
-
-
 <h2>Other Rooms</h2>
 
-<ul> 
 {{#each model.orphan_rooms as |r|}}
-  <li>{{#link-to 'location' r.id}}{{r.name}}{{/link-to}}</li>
+<div class="panel panel-default">
+   <div class="panel-heading">{{r.name}}</div>
+   <div class="panel-body">{{{ansi-format text=r.description}}}</div>
+</div>
 {{/each}}
-</ul>
 
-{{#bs-modal-simple
-    title="Confirm Delete"
-    closeTitle="Cancel"
-    submitTitle="Confirm"
-    size=null
-    fade=false
-    open=confirmDelete
-    onSubmit=(action "delete")
-    onHide=(action (mut confirmDelete) null)}}
-    You're about to delete area {{currentArea.name}}. This cannot be undone.  Are you sure?
-  {{/bs-modal-simple}}
-  
 {{outlet}}
