+++ app/templates/wiki-edit.hbs
@@ -35,10 +35,10 @@
     
     {{#if preview}}
     <div class="panel panel-default">
-        <div class="panel panel-heading">
+        <div class="panel-heading">
             Preview
         </div>
-        <div class="panel panel-body">
+        <div class="panel-body">
             {{{preview}}}
         </div>
     </div>
